+++ R/globals.R
@@ -1,5 +1,1 @@
-<<<<<<< HEAD
-globalVariables(c('prov_code','1999_nickname','.'))
-=======
 globalVariables(c('prov_code','1999_nickname','.','dia_sub_group','freq', 'prov_scode'))
->>>>>>> 5d365d016fdaf9f619a0ac22d311ef697e9adb5e